net: SocketWrapper: Refactoring
split header files in header + sources
