+++ libraries/SocketWrapper/ZephyrClient.h
@@ -9,107 +9,36 @@
 #include "SocketWrapper.h"
 #include "api/Client.h"
 #include "unistd.h"
-#include "zephyr/sys/printk.h"
-#include <errno.h>
 
 class ZephyrClient : public arduino::Client, ZephyrSocketWrapper {
 private:
 	bool _connected = false;
 
 protected:
-	void setSocket(int sock) {
-		sock_fd = std::shared_ptr<int>(sock < 0 ? nullptr : new int(sock), socket_deleter);
-		_connected = true;
-	}
+	void setSocket(int sock) override;
 
 public:
-	int connect(const char *host, uint16_t port) override {
-		auto ret = ZephyrSocketWrapper::connect((char *)host, port);
-		if (ret) {
-			_connected = true;
-		}
-		return ret;
-	}
+	int connect(const char *host, uint16_t port) override;
+	int connect(IPAddress ip, uint16_t port);
 
-	int connect(IPAddress ip, uint16_t port) {
-		auto ret = ZephyrSocketWrapper::connect(ip, port);
-		if (ret) {
-			_connected = true;
-		}
-		return ret;
-	}
 #if defined(CONFIG_NET_SOCKETS_SOCKOPT_TLS)
-	int connectSSL(const char *host, uint16_t port, const char *cert) {
-		auto ret = ZephyrSocketWrapper::connectSSL(host, port, cert);
-		if (ret) {
-			_connected = true;
-		}
-		return ret;
-	}
+	int connectSSL(const char *host, uint16_t port, const char *cert);
 #endif
-	uint8_t connected() override {
-		uint8_t buf;
-		int ret = ::recv(*sock_fd, &buf, 1, MSG_PEEK | MSG_DONTWAIT);
-		if (ret == 0) {
-			stop();
-		}
-		return _connected;
-	}
+	uint8_t connected() override;
 
-	int available() override {
-		return ZephyrSocketWrapper::available();
-	}
+	int available() override;
+	int read() override;
+	int read(uint8_t *buffer, size_t size) override;
+	size_t write(uint8_t c) override;
+	size_t write(const uint8_t *buffer, size_t size) override;
 
-	int read() override {
-		uint8_t c;
-		read(&c, 1);
-		return c;
-	}
+	void flush() override;
 
-	int read(uint8_t *buffer, size_t size) override {
-		auto received = recv(buffer, size);
+	int peek() override;
+	void stop() override;
 
-		if (received > 0) {
-			return received;
-		}
-		if (received == 0) {
-			return 0;
-		}
-		if (errno == EAGAIN || errno == EWOULDBLOCK) {
-			return 0;
-		}
-		return -1;
-	}
+	operator bool();
 
-	size_t write(uint8_t c) override {
-		return write(&c, 1);
-	}
-
-	size_t write(const uint8_t *buffer, size_t size) override {
-		return send(buffer, size);
-	}
-
-	void flush() override {
-		// No-op
-	}
-
-	int peek() override {
-		uint8_t c;
-		recv(&c, 1, MSG_PEEK);
-		return c;
-	}
-
-	void stop() override {
-		ZephyrSocketWrapper::close();
-		_connected = false;
-	}
-
-	operator bool() {
-		return (sock_fd != nullptr);
-	}
-
-	String remoteIP() {
-		return ZephyrSocketWrapper::remoteIP();
-	}
+	String remoteIP();
 	friend class ZephyrServer;
 };
