+++ libraries/SocketWrapper/ZephyrServer.h
@@ -16,55 +16,23 @@ private:
 	int _port;
 
 public:
-	ZephyrServer() : _port(80) {};
-	ZephyrServer(uint16_t port) : _port(port) {};
+	ZephyrServer();
+	ZephyrServer(uint16_t port);
 
-	virtual ~ZephyrServer() {
-		end();
-	}
+	virtual ~ZephyrServer();
 
-	void end() {
-		ZephyrSocketWrapper::close();
-	}
+	void end();
+	void begin(uint16_t port);
+	void begin();
 
-	void begin(uint16_t port) {
-		_port = port;
-		begin();
-	}
+	uint8_t status();
 
-	void begin() {
-		ZephyrSocketWrapper::bind(_port);
-		ZephyrSocketWrapper::listen(5);
-	}
+	explicit operator bool();
 
-	uint8_t status() {
-		return 0;
-	}
-
-	explicit operator bool() {
-		return (sock_fd != nullptr);
-	}
-
-	ZephyrClient accept(uint8_t *status = nullptr) {
-		ARG_UNUSED(status);
-
-		ZephyrClient client;
-		int sock = ZephyrSocketWrapper::accept();
-		client.setSocket(sock);
-		return client;
-	}
-
-	ZephyrClient available(uint8_t *status = nullptr) __attribute__((deprecated("Use accept()."))) {
-		return accept(status);
-	}
-
-	size_t write(uint8_t c) override {
-		return write(&c, 1);
-	}
-
-	size_t write(const uint8_t *buffer, size_t size) override {
-		return send(buffer, size);
-	}
+	ZephyrClient accept(uint8_t *status = nullptr);
+	ZephyrClient available(uint8_t *status = nullptr) __attribute__((deprecated("Use accept().")));
+	size_t write(uint8_t c) override;
+	size_t write(const uint8_t *buffer, size_t size) override;
 
 	friend class ZephyrClient;
 };
